Move uploading logic to Review
